Extracted baseclass for the application
